+++ src/containers/app/index.js
@@ -1,0 +1,5 @@
+import App from './App'
+
+export default {
+  App,
+}
