+++ src/containers/index.js
@@ -1,0 +1,5 @@
+import app  from './app';
+
+export {
+  app
+}
