+++ src/index.js
@@ -1,0 +1,12 @@
+import React from 'react'
+import { Component } from 'react-native';
+
+import { app } from './containers';
+
+const Main = () => {
+  return (
+      <app.App />
+  )
+}
+
+export default Main
